requirements.txt
requests 
kafka-python 
cassandra-driver